Ensure no panic in goroutine
In Go, applications typically recover from panic in the request
goroutine. But there is no way to recover from a gorouting
spawned in the driver. If the code panics, it will case the
process to shutdown (or server restart in prod). This is not good.
To prevent this from happening, we always defer a recover inside
the gorouting function.

Signed-off-by: Andre Furlan <andre.furlan@databricks.com>